refactor: move RunnerArtifactRef to runner-contract
Moves the plain-data RunnerArtifactRef (artifact_id + optional name/path/url/
mime/size/sha256/transport) into homeboy-runner-contract so core can name it
without a core -> runner edge. The From<&JobArtifactMetadata> impl stays in
core (orphan rule: JobArtifactMetadata is core-owned in api_jobs/remote_runner).

Re-exported from runner::session so internal/CLI call sites resolve unchanged;
external consumers (observation/runs_service, run_outcome_envelope) repointed to
homeboy_runner_contract::.

This also unblocks the upcoming runner-evidence hook: download_remote_artifact
returns a RemoteArtifactDownload whose artifact_ref is a RunnerArtifactRef, and
ArtifactFetchOutcome carries Option<RunnerArtifactRef> — both now nameable in
core without a runner edge.

Part of the staged runner extraction plan (wiki: homeboy-runner-crate-
extraction-plan; hook design: runner-evidence-hook-design).

Verified: cargo check clean for -p homeboy-runner-contract, -p homeboy-core
--lib, -p homeboy-cli --lib.
