+++ crates/homeboy-core/src/run_outcome_envelope.rs
@@ -3,8 +3,9 @@ use serde_json::Value;
 
 use crate::api_jobs::JobArtifactMetadata;
 use crate::artifact_ref::{artifact_uri, ArtifactRef, EvidenceRef, ARTIFACT_REF_SCHEMA};
-use crate::runner::{RunnerArtifactRef, RunnerHandoff};
+use crate::runner::RunnerHandoff;
 use crate::runner_execution_envelope::{RunnerExecutionArtifactRef, RunnerExecutionRecord};
+use homeboy_runner_contract::RunnerArtifactRef;
 
 pub const RUN_OUTCOME_ENVELOPE_SCHEMA: &str = "homeboy/run-outcome-envelope/v1";
 pub const RUN_OUTCOME_ENVELOPE_FILE: &str = "run-outcome-envelope.json";
